mitigation: add fallback resources
It has been observed, that mitigations would fail in certain situations.
The specific situation observed was that writing invalid paramters to the cpuquiet sysfs resource nr_thermal_max_cpus.
The kernel would refuse such writes, rendering the mitigation applied nonfunctional.
This can be dangerous when the mitigation is critical for device health.
In order to make sure mitigations always apply, this commit introduces a basic fallback implementation.
An example of a mitigation with a fallback resource:
<mitigation level="1">
  <value resource="normal-resource">1</value>
  <value fallback="fallback" resource="fallback-resource">1</value>
</mitigation>
If the write of normal-resource fails, then thermanager will write to fallback-resource.

Change-Id: Iad8eae562b3627f9a6ef66bbca35d66c202542b3